Added Texture, Game, Rendering impl
I added a texture class, which can be rendered (not tested) from a renderer. I added a game.hpp/cpp file(s), which are not fully implemented but get the general idea down. They initialize, run, loop, and clean resources at the end of the game loop. Rendering files are declared, but not implemented yet.